+++ python/ray/workflow/tests/test_virtual_actor_3.py
@@ -1,8 +1,11 @@
 import pytest
+from filelock import FileLock
+import time
+from typing import Optional, Dict, Tuple, List
 
 import ray
 from ray import workflow
-from typing import Optional, Dict, Tuple, List
+from ray.tests.conftest import *  # noqa
 
 
 @ray.workflow.virtual_actor
@@ -121,6 +124,86 @@ def test_writer_actor_pressure_test(workflow_start_regular):
     assert ray.get(balance_3) == 80
     assert user.balance.run() == 85
     assert user.goods_value.run() == (5, [])
+
+
+@pytest.mark.parametrize(
+    "workflow_start_regular",
+    [
+        {
+            "num_cpus": 4
+            # We need more CPUs, otherwise 'create()' blocks 'get()'
+        }
+    ],
+    indirect=True,
+)
+@pytest.mark.repeat(5)
+def test_wf_in_actor_chain(workflow_start_regular, tmp_path):
+    file_lock = [str(tmp_path / str(i)) for i in range(5)]
+    fail_flag = tmp_path / "fail"
+
+    @workflow.virtual_actor
+    class Counter:
+        def __init__(self):
+            self._counter = 0
+
+        def incr(self, n):
+            with FileLock(file_lock[n]):
+                self._counter += 1
+                if fail_flag.exists():
+                    raise Exception()
+
+            if n == 0:
+                return self._counter
+            else:
+                return self.incr.step(n - 1)
+
+        @workflow.virtual_actor.readonly
+        def val(self):
+            return self._counter
+
+        def __getstate__(self):
+            return self._counter
+
+        def __setstate__(self, v):
+            self._counter = v
+
+    locks = [FileLock(f) for f in file_lock]
+    for lock in locks:
+        lock.acquire()
+
+    c = Counter.get_or_create("counter")
+    ray.get(c.ready())
+    final_ret = c.incr.run_async(len(file_lock) - 1)
+    for i in range(0, len(file_lock) - 2):
+        locks[-i - 1].release()
+        val = c.val.run()
+        for _ in range(0, 60):
+            if val == i + 1:
+                break
+            val = c.val.run()
+            time.sleep(1)
+        assert val == i + 1
+
+    fail_flag.touch()
+    locks[1 - len(file_lock)].release()
+    # Fail the pipeline
+    with pytest.raises(Exception):
+        ray.get(final_ret)
+
+    fail_flag.unlink()
+    workflow.resume("counter")
+    # After resume, it'll start form the place where it failed
+    for i in range(len(file_lock) - 1, len(file_lock)):
+        locks[-i - 1].release()
+        val = c.val.run()
+        for _ in range(0, 60):
+            if val == i + 1:
+                break
+            val = c.val.run()
+            time.sleep(1)
+        assert val == i + 1
+
+    assert c.val.run() == 5
 
 
 if __name__ == "__main__":
